Updated texts. Changed subject duration from int 2 to tinyint 1. Changed processing to treat the duration as an int after the initial string extraction. Semester(s) is now added as a unit to the duration output.

+++ com_organizer/admin/sql/install.mysql.utf8.sql
@@ -818,7 +818,7 @@ CREATE TABLE IF NOT EXISTS `#__organizer_subjects` (
     `creditPoints`                INT(3) UNSIGNED  NOT NULL DEFAULT 0,
     `description_de`              TEXT,
     `description_en`              TEXT,
-    `duration`                    INT(2) UNSIGNED           DEFAULT 1,
+    `duration`                    TINYINT(1) UNSIGNED       DEFAULT 1,
     `expenditure`                 INT(4) UNSIGNED  NOT NULL DEFAULT 0,
     `expertise`                   TINYINT(1) UNSIGNED       DEFAULT NULL,
     `expertise_de`                TEXT,
